+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Disallow negative depth values.
+- Use the `AbortSignal.throwIfAborted` method.
+
 ## [0.5.1] - 2025-12-15
 
 - Improve README targets section and jsdoc comments.